+++ Addon/ElvUI_mMediaTag/core/changelog/3.62.lua
@@ -1,0 +1,7 @@
+mMT.Changelog[362]  = {
+	DATE = "14.10.2024",
+	FIX = {
+		"[Objectivtracker]: Skin for Delves and Dungeons.",
+		"[Objectivtracker]: Removed Test functions.",
+	},
+}