+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog - ElvUI_mMediaTag
 [Eng] - All changes to this project will be documented in this file. The latest changes are at the top.
 
+## [ver. 3.62] - 14.10.2024
+### FIX 
+- FIX - [Objectivtracker]: Skin for Delves and Dungeons.
+- FIX - [Objectivtracker]: Removed Test functions.
+
 ## [ver. 3.61] - 13.10.2024
 ### UPDATE
 - UPDATE - [Datatext-Score]: Update rewards infos.